refactor: Move VerneedTable to elf.rs
